Add files via upload
Uri and I 98% completed with the retrieving images for the database connection class. Validation.java and Registration.java each have one new line that sets the attribute of the new user that registers or logs in. Fixing up CollageDisplay.jsp for the next commit and then Save To History is set and done.